feat: refactored common WeakMapIterator logic
