+++ src/ros/rover/teleop_arm_joy/include/teleop_arm_joy/inputs/InputCollection.hpp
@@ -36,132 +36,8 @@ public:
   InputCollection& operator=(const InputCollection&) = delete;
 
   // Container type aliases
-  using value_type = std::shared_ptr<InputT>;
-  using reference = value_type&;
-  using const_reference = const value_type&;
-  using size_type = std::size_t;
-  using difference_type = std::ptrdiff_t;
-
-  template<bool is_const>
-  class Iterator {
-  public:
-    using iterator_category = std::bidirectional_iterator_tag;
-    using value_type = std::shared_ptr<InputT>;
-    using difference_type = std::ptrdiff_t;
-
-    // Conditional types based on const-ness
-    using reference = typename std::conditional_t<is_const,
-      const value_type&,
-      value_type&>;
-    using pointer = typename std::conditional_t<is_const,
-      const value_type*,
-      value_type*>;
-    using base_iterator = typename std::conditional_t<is_const,
-      typename std::map<std::string, std::weak_ptr<InputT>>::const_iterator,
-      typename std::map<std::string, std::weak_ptr<InputT>>::iterator>;
-
-    using map_type = std::map<std::string, std::weak_ptr<InputT>>;
-    using map_pointer = typename std::conditional_t<is_const,
-      const map_type*,
-      map_type*>;
-
-    explicit Iterator(base_iterator it, map_pointer map)
-      : it_(it), map_(map), current_value_()
-    {
-      if (it_ != map_->end()) {
-        skip_expired();
-      }
-    }
-
-
-    // Convert non-const iterator to const iterator
-    template<bool was_const, typename = std::enable_if_t<is_const && !was_const>>
-    explicit Iterator(const Iterator<was_const>& other)
-      : it_(other.it_), map_(other.map_), current_value_(other.current_value_) {}
-
-    reference operator*() {
-      ensure_current_value();
-      return current_value_;
-    }
-
-    pointer operator->() {
-      ensure_current_value();
-      return &current_value_;
-    }
-
-    // Prefix increment
-    Iterator& operator++() {
-      ++it_;
-      skip_expired();
-      return *this;
-    }
-
-    // Postfix increment
-    Iterator operator++(int) {
-      Iterator tmp = *this;
-      ++(*this);
-      return tmp;
-    }
-
-    friend bool operator==(const Iterator& a, const Iterator& b) {
-      return a.it_ == b.it_;
-    }
-
-    friend bool operator!=(const Iterator& a, const Iterator& b) {
-      return !(a == b);
-    }
-
-    /// Allows access to the base iterator for conversion
-    base_iterator base() const { return it_; }
-
-  private:
-    base_iterator it_;
-    map_pointer map_;
-    value_type current_value_;
-
-    /// Skips over any weak pointers that have expired
-    void skip_expired() {
-      while (it_ != map_->end()) {
-        if (auto ptr = it_->second.lock()) {
-          current_value_ = ptr;
-          return;
-        }
-
-        // Remove any expired entry before continuing to the next item
-        if constexpr (!is_const) {
-          it_ = map_->erase(it_);
-        } else {
-          ++it_;
-        }
-      }
-
-      current_value_.reset();
-    }
-
-    void ensure_current_value() {
-      if (current_value_)
-        return;
-
-      if (auto ptr = it_->second.lock()) {
-        current_value_ = ptr;
-        return;
-      }
-
-      // If current value expired, skip to next valid one
-      if constexpr (!is_const) {
-        it_ = map_->erase(it_);
-      } else {
-        ++it_;
-      }
-      skip_expired();
-    }
-
-    // Grant access to other iterator specializations
-    template<bool> friend class Iterator;
-  };
-
-  using iterator = Iterator<false>;
-  using const_iterator = Iterator<true>;
+  using iterator = WeakMapIterator<InputT, false>;
+  using const_iterator = WeakMapIterator<InputT, true>;
 
   std::shared_ptr<InputT> operator[](const std::string& key) {
     // Find the element
