+++ src/ros/rover/teleop_arm_joy/include/teleop_arm_joy/inputs/events/EventCollection.hpp
@@ -10,6 +10,7 @@
 #include <memory>
 #include <map>
 #include "Event.hpp"
+#include "../WeakMapIterator.hpp"
 
 namespace teleop_arm_joy {
 
@@ -19,110 +20,10 @@ namespace teleop_arm_joy {
 class EventCollection {
 public:
   EventCollection() = default;
-  explicit EventCollection(std::weak_ptr<EventListenerQueue> listener_queue)
-    : listener_queue_(std::move(listener_queue)) {}
+  explicit EventCollection(std::weak_ptr<EventListenerQueue> listener_queue);
 
-  template<bool is_const>
-  class Iterator {
-  public:
-    using iterator_category = std::bidirectional_iterator_tag;
-    using value_type = Event::SharedPtr;
-    using difference_type = std::ptrdiff_t;
-
-    using reference = typename std::conditional_t<is_const,
-      const value_type&,
-      value_type&>;
-    using pointer = typename std::conditional_t<is_const,
-      const value_type*,
-      value_type*>;
-    using base_iterator = typename std::conditional_t<is_const,
-      typename std::map<std::string, std::weak_ptr<Event>>::const_iterator,
-      typename std::map<std::string, std::weak_ptr<Event>>::iterator>;
-
-    using map_type = std::map<std::string, std::weak_ptr<Event>>;
-    using map_pointer = typename std::conditional_t<is_const,
-      const map_type*,
-      map_type*>;
-
-    explicit Iterator(base_iterator it, map_pointer map)
-      : it_(it), map_(map), current_value_()
-    {
-      if (it_ != map_->end()) {
-        skip_expired();
-      }
-    }
-
-    reference operator*() {
-      ensure_current_value();
-      return current_value_;
-    }
-
-    pointer operator->() {
-      ensure_current_value();
-      return &current_value_;
-    }
-
-    Iterator& operator++() {
-      ++it_;
-      skip_expired();
-      return *this;
-    }
-
-    Iterator operator++(int) {
-      Iterator tmp = *this;
-      ++(*this);
-      return tmp;
-    }
-
-    friend bool operator==(const Iterator& a, const Iterator& b) {
-      return a.it_ == b.it_;
-    }
-
-    friend bool operator!=(const Iterator& a, const Iterator& b) {
-      return !(a == b);
-    }
-
-  private:
-    base_iterator it_;
-    map_pointer map_;
-    value_type current_value_;
-
-    void skip_expired() {
-      while (it_ != map_->end()) {
-        if (auto ptr = it_->second.lock()) {
-          current_value_ = ptr;
-          return;
-        }
-
-        if constexpr (!is_const) {
-          it_ = map_->erase(it_);
-        } else {
-          ++it_;
-        }
-      }
-      current_value_.reset();
-    }
-
-    void ensure_current_value() {
-      if (current_value_)
-        return;
-
-      if (auto ptr = it_->second.lock()) {
-        current_value_ = ptr;
-        return;
-      }
-
-      if constexpr (!is_const) {
-        it_ = map_->erase(it_);
-      } else {
-        ++it_;
-      }
-      skip_expired();
-    }
-  };
-
-  using iterator = Iterator<false>;
-  using const_iterator = Iterator<true>;
+  using iterator = WeakMapIterator<Event, false>;
+  using const_iterator = WeakMapIterator<Event, true>;
 
   Event::SharedPtr operator[](const std::string& index) {
     auto it = items_.find(index);
@@ -150,28 +51,12 @@ public:
   const_iterator cbegin() const { return const_iterator(items_.begin(), &items_); }
   const_iterator cend() const { return const_iterator(items_.end(), &items_); }
 
-  void clean_up() {
-    for (auto it = items_.begin(); it != items_.end();) {
-      if (it->second.expired()) {
-        it = items_.erase(it);
-      } else {
-        ++it;
-      }
-    }
-  }
+  void clean_up();
 
   /**
    * Gets the number of non-expired inputs
    */
-  [[nodiscard]] size_t size() const {
-    size_t count = 0;
-    for (const auto& item : items_) {
-      if (!item.second.expired()) {
-        ++count;
-      }
-    }
-    return count;
-  }
+  [[nodiscard]] size_t size() const;
 
 private:
   std::map<std::string, std::weak_ptr<Event>> items_{};
